🎨 feat: 판매자 문의 타입에러 해결
