+++ src/types/shopping_detail.ts
@@ -63,3 +63,30 @@ export interface ShoppingFormType {
   reviewCount: number;
   avg: number;
 }
+
+// 사용자 정보 타입
+export interface User {
+  _id: number;
+  name: string;
+  image: string;
+}
+// 상품 정보 타입
+export interface Product {
+  name: string;
+}
+// 메인 문의 타입
+export interface InquiryItem {
+  _id: number;
+  type: 'qna';
+  product_id: number;
+  seller_id: number;
+  views: number;
+  user: User;
+  title: string;
+  content: string;
+  createdAt: string;
+  updatedAt: string;
+  product: Product;
+  bookmarks: number;
+  repliesCount: number;
+}
